move conftest back to worker
